# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JST Tailors Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
            description: "Date in YYYY-MM-DD format, defaults to today"
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly report
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, customerId, orderDate, status, totalAmount, advanceAmount, balanceAmount]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderDate:
          type: string
        deliveryDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_stitching, ready, delivered]
        totalAmount:
          type: number
        advanceAmount:
          type: number
        balanceAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Payment:
      type: object
      required: [id, orderId, customerId, amount, paymentDate]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        orderNumber:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DailyReport:
      type: object
      required: [date, newOrders, deliveredOrders, totalCollection, newCustomers]
      properties:
        date:
          type: string
        newOrders:
          type: integer
        deliveredOrders:
          type: integer
        totalCollection:
          type: number
        newCustomers:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    MonthlyReport:
      type: object
      required: [year, month, totalOrders, totalRevenue, totalCollected, totalBalance, newCustomers]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        totalCollected:
          type: number
        totalBalance:
          type: number
        newCustomers:
          type: integer
        ordersByStatus:
          type: object
          properties:
            pending:
              type: integer
            in_stitching:
              type: integer
            ready:
              type: integer
            delivered:
              type: integer
...