# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JST Tailors Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';
import type { Payment } from './payment';

export interface DailyReport {
  date: string;
  newOrders: number;
  deliveredOrders: number;
  totalCollection: number;
  newCustomers: number;
  orders?: Order[];
  payments?: Payment[];
}